+++ src/HttpClient/Builder.php
@@ -114,7 +114,7 @@ final class Builder
     /**
      * @return HttpMethodsClientInterface
      */
-    public function getHttpClient()
+    public function getHttpClient(): HttpMethodsClientInterface
     {
         if (null === $this->pluginClient) {
             $plugins = $this->plugins;
@@ -137,7 +137,7 @@ final class Builder
      *
      * @return RequestFactoryInterface
      */
-    public function getRequestFactory()
+    public function getRequestFactory(): RequestFactoryInterface
     {
         return $this->requestFactory;
     }
@@ -147,7 +147,7 @@ final class Builder
      *
      * @return StreamFactoryInterface
      */
-    public function getStreamFactory()
+    public function getStreamFactory(): StreamFactoryInterface
     {
         return $this->streamFactory;
     }
@@ -157,7 +157,7 @@ final class Builder
      *
      * @return UriFactoryInterface
      */
-    public function getUriFactory()
+    public function getUriFactory(): UriFactoryInterface
     {
         return $this->uriFactory;
     }