+++ src/HttpClient/Util/JsonArray.php
@@ -30,7 +30,7 @@ final class JsonArray
      *
      * @return array
      */
-    public static function decode(string $json)
+    public static function decode(string $json): array
     {
         /** @var scalar|array|null */
         $data = \json_decode($json, true);
@@ -55,7 +55,7 @@ final class JsonArray
      *
      * @return string
      */
-    public static function encode(array $value)
+    public static function encode(array $value): string
     {
         $json = \json_encode($value);