+++ src/ResultPagerInterface.php
@@ -15,6 +15,7 @@ declare(strict_types=1);
 namespace Gitlab;
 
 use Gitlab\Api\AbstractApi;
+use Generator;
 
 /**
  * This is the result pager interface.
@@ -36,7 +37,7 @@ interface ResultPagerInterface
      *
      * @return array
      */
-    public function fetch(AbstractApi $api, string $method, array $parameters = []);
+    public function fetch(AbstractApi $api, string $method, array $parameters = []): array;
 
     /**
      * Fetch all results from an api call.
@@ -49,7 +50,7 @@ interface ResultPagerInterface
      *
      * @return array
      */
-    public function fetchAll(AbstractApi $api, string $method, array $parameters = []);
+    public function fetchAll(AbstractApi $api, string $method, array $parameters = []): array;
 
     /**
      * Lazily fetch all results from an api call.
@@ -62,14 +63,14 @@ interface ResultPagerInterface
      *
      * @return \Generator
      */
-    public function fetchAllLazy(AbstractApi $api, string $method, array $parameters = []);
+    public function fetchAllLazy(AbstractApi $api, string $method, array $parameters = []): Generator;
 
     /**
      * Check to determine the availability of a next page.
      *
      * @return bool
      */
-    public function hasNext();
+    public function hasNext(): bool;
 
     /**
      * Fetch the next page.
@@ -78,14 +79,14 @@ interface ResultPagerInterface
      *
      * @return array
      */
-    public function fetchNext();
+    public function fetchNext(): array;
 
     /**
      * Check to determine the availability of a previous page.
      *
      * @return bool
      */
-    public function hasPrevious();
+    public function hasPrevious(): bool;
 
     /**
      * Fetch the previous page.
@@ -94,7 +95,7 @@ interface ResultPagerInterface
      *
      * @return array
      */
-    public function fetchPrevious();
+    public function fetchPrevious(): array;
 
     /**
      * Fetch the first page.
@@ -103,7 +104,7 @@ interface ResultPagerInterface
      *
      * @return array
      */
-    public function fetchFirst();
+    public function fetchFirst(): array;
 
     /**
      * Fetch the last page.
@@ -112,5 +113,5 @@ interface ResultPagerInterface
      *
      * @return array
      */
-    public function fetchLast();
+    public function fetchLast(): array;
 }